requirements.txt
flask
mysql-connector-python
pycryptodome
